+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+## Version 0.1.5
+
+Adding class method for accessing default asset definitions
+
+## Version 0.1.4
+
+- Modify asset definitions array returned form get_asset_defs()
+- Set logger default level to info
+
+## Version 0.1.3
+
+Modify assets array returned from get_assets()
+
+## Version 0.1.2
+
+Fix access to default asset_definitions.json file
+
 ## Version 0.1.1
 
 Changes to output structure and schema